Consolidate color logic
