--- packages/components/src/List/utils/listItemLabelColor.ts
@@ -1,42 +1,0 @@
-/*
-
- MIT License
-
- Copyright (c) 2021 Looker Data Sciences, Inc.
-
- Permission is hereby granted, free of charge, to any person obtaining a copy
- of this software and associated documentation files (the "Software"), to deal
- in the Software without restriction, including without limitation the rights
- to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- copies of the Software, and to permit persons to whom the Software is
- furnished to do so, subject to the following conditions:
-
- The above copyright notice and this permission notice shall be included in all
- copies or substantial portions of the Software.
-
- THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- SOFTWARE.
-
- */
-
-import { ListColor, listItemColorAppliesToLabel, listItemColor } from '../types'
-
-export const listItemLabelColor = (color?: ListColor, disabled?: boolean) => {
-  if (disabled) {
-    return 'text1'
-  } else if (color) {
-    if (listItemColorAppliesToLabel.includes(color)) {
-      // Theme Color & color is applied to label
-      return color
-    } else if (!listItemColor.includes(color)) {
-      // HTML color
-      return color
-    }
-  }
-  return 'text5'
-}
